Добавила модуляцию
