+++ index.html
@@ -13,7 +13,7 @@
   <main>
 <section class="function">
 <!--  <script src="js/function.js"></script>-->
-  <script src="js/main.js"></script>
+  <script src="js/main.js" type="module"></script>
 </section>
     <!-- Фильтрация изображений от других пользователей -->
     <section class="img-filters  img-filters--inactive  container">
