+++ js/data-storage.js
@@ -1,0 +1,62 @@
+const DESCRIPTION = [
+  'Тяжелый день на болоте. Заказ пиццы – единственная активность на вечер.',
+  'Этот вазон сам бросился под меня. Я лишь попытался его спасти.',
+  'Моё лицо, когда ты в третий раз рассказываешь эту историю.',
+  'Моментальный снимок моей продуктивности. Она только что была здесь.',
+  'Мама, можно мы возьмём того блестящего жука? – Мама: У нас дома есть блестящие жуки. Блестящие жуки дома: пачка скрепок.',
+  'План был безупречен. Исполнение – не очень. Но печенье того стоило.',
+  'Когда пытаешься эффектно уйти, но спотыкаешься о собственную крутость.',
+  'Окак.',
+  'Моё отношение к вашему ЗОЖ. Где мои макароны с сыром?',
+  'Енот ушёл в себя. Вернётся с добычей.',
+  'Енот видит еду. Енот хочет еду. Енот применяет стратегию «увидел-украл-убежал».',
+  'Я не будильник, чтобы квакать по утрам',
+  'Лягушечковое настроение: сонное'
+];
+
+const COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const NAMES = [
+  'Василий',
+  'Иннокентий',
+  'Семён',
+  'Акакий',
+  'Федот',
+  'Кузьма',
+  'Порфирий',
+  'Евлампий',
+  'Прокофий',
+  'Захар',
+  'Афанасий',
+  'Терентий',
+  'Спиридон',
+  'Архип',
+  'Геннадий'
+];
+
+const SURNAMES = [
+  'Пукин',
+  'Хрюндель',
+  'Помидоров',
+  'Шпунтиков',
+  'Кастрюлькин',
+  'Шпроткин',
+  'Тапкоходов',
+  'Воробушкин',
+  'Картофелин',
+  'Мухоморов',
+  'Пузырёв',
+  'Чайников',
+  'Блинчиков',
+  'Огурцов',
+  'Тапочкин'
+];
+
+export {DESCRIPTION, COMMENTS, NAMES, SURNAMES};
